Removed dependency for email verification
Removed the dependency for email verification through Resend. This is because Resend requires a custom domain to be able to send emails to registered users. This also makes it easier for users to test the web application since they were unable to test it properly if they hadn't verified their email.

If a custom email domain is at hand, apply the "RESEND_EMAIL" .env variable.